Acts as Resource Controller Module
Move core of ResourceController to a module, ActsAsResourceController.
