+++ lib/jsonapi-resources.rb
@@ -1,4 +1,5 @@
 require 'jsonapi/resource'
+require 'jsonapi/acts_as_resource_controller'
 require 'jsonapi/resource_controller'
 require 'jsonapi/resources/version'
 require 'jsonapi/configuration'
